# Fixer: improve performance of generateDiff()