--- a/tests/Core/Fixer/Fixtures/GenerateDiffTest-BlankLinesAtEnd.inc
+++ b/tests/Core/Fixer/Fixtures/GenerateDiffTest-BlankLinesAtEnd.inc
@@ -1,0 +1,11 @@
+<?php
+// Comment with 2 spaces trailing whitespace.  
+$var = FALSE;
+
+if ($var) {
+	echo 'This line is tab indented';
+}
+
+
+
+