--- a/tests/Core/Fixer/Fixtures/GenerateDiffTest-BlankLinesAtStart.inc
+++ b/tests/Core/Fixer/Fixtures/GenerateDiffTest-BlankLinesAtStart.inc
@@ -1,0 +1,10 @@
+
+
+
+<?php
+// Comment with 2 spaces trailing whitespace.  
+$var = FALSE;
+
+if ($var) {
+	echo 'This line is tab indented';
+}