--- a/tests/Core/Fixer/Fixtures/GenerateDiffTest-LineAdded.diff
+++ b/tests/Core/Fixer/Fixtures/GenerateDiffTest-LineAdded.diff
@@ -1,0 +1,8 @@
+--- tests/Core/Fixer/Fixtures/GenerateDiffTest-LineAdded.inc
++++ PHP_CodeSniffer
+@@ -1,4 +1,5 @@
+ <?php
++// Comment with 2 spaces trailing whitespace.  
+ $var = FALSE;
+ 
+ if ($var) {