--- a/tests/Core/Fixer/Fixtures/GenerateDiffTest-LineAdded.inc
+++ b/tests/Core/Fixer/Fixtures/GenerateDiffTest-LineAdded.inc
@@ -1,0 +1,6 @@
+<?php
+$var = FALSE;
+
+if ($var) {
+	echo 'This line is tab indented';
+}