--- a/tests/Core/Fixer/Fixtures/GenerateDiffTest-LineRemoved.inc
+++ b/tests/Core/Fixer/Fixtures/GenerateDiffTest-LineRemoved.inc
@@ -1,0 +1,8 @@
+<?php
+// Comment with 2 spaces trailing whitespace.  
+$var = FALSE;
+
+if ($var) {
+	echo 'This line is tab indented';
+    echo 'And this line is not';
+}