--- a/tests/Core/Fixer/Fixtures/GenerateDiffTest-NoTrailingWhitespace.diff
+++ b/tests/Core/Fixer/Fixtures/GenerateDiffTest-NoTrailingWhitespace.diff
@@ -1,0 +1,9 @@
+--- tests/Core/Fixer/Fixtures/GenerateDiffTest-NoTrailingWhitespace.inc
++++ PHP_CodeSniffer
+@@ -1,5 +1,5 @@
+ <?php
+-// Comment with 2 spaces trailing whitespace.
++// Comment with 2 spaces trailing whitespace.  
+ $var = FALSE;
+ 
+ if ($var) {