--- a/tests/Core/Fixer/Fixtures/GenerateDiffTest-VarNameChanged.inc
+++ b/tests/Core/Fixer/Fixtures/GenerateDiffTest-VarNameChanged.inc
@@ -1,0 +1,7 @@
+<?php
+// Comment with 2 spaces trailing whitespace.  
+$rav = FALSE;
+
+if ($rav) {
+	echo 'This line is tab indented';
+}