--- a/tests/Core/Fixer/Fixtures/GenerateDiffTest-WhiteSpaceAtStart.diff
+++ b/tests/Core/Fixer/Fixtures/GenerateDiffTest-WhiteSpaceAtStart.diff
@@ -1,0 +1,8 @@
+--- tests/Core/Fixer/Fixtures/GenerateDiffTest-WhiteSpaceAtStart.inc
++++ PHP_CodeSniffer
+@@ -1,4 +1,4 @@
+- <?php
++<?php
+ // Comment with 2 spaces trailing whitespace.  
+ $var = FALSE;
+