--- a/tests/Core/Fixer/Fixtures/GenerateDiffTest.inc
+++ b/tests/Core/Fixer/Fixtures/GenerateDiffTest.inc
@@ -1,0 +1,7 @@
+<?php
+// Comment with 2 spaces trailing whitespace.  
+$var = FALSE;
+
+if ($var) {
+	echo 'This line is tab indented';
+}