@jrfnl 2024-02-21
I've updated the first commit with four extra tests (diff of blank lines/whitespace at start/end of file)