@rodrigoprimo 2024-02-21
> Based on the test runs in CI, this change should make no significant different on *nix systems, but I'd love to have some people test it and post their with and without timing results.

I tested this on my Ubuntu machine using the command `vendor/bin/phpunit tests/AllTests.php --no-coverage`, and I saw a small increase in the execution time of the tests. I did only a few runs on the main branch and on this branch. On the main branch, it takes around 26.5 seconds for the command to finish. On this branch, it takes around 27.5 seconds. I would be happy to check this further if you would like me to.

On a quick check, I believe the same happened in the CI runs. On the last few runs that I checked on the main branch, it took around 13 seconds to run the tests (https://github.com/PHPCSStandards/PHP_CodeSniffer/actions/runs/7977678082/job/21781108952#step:9:56 and https://github.com/PHPCSStandards/PHP_CodeSniffer/actions/runs/7975902826/job/21775216017#step:9:56). On this branch, it took around 15 seconds (https://github.com/PHPCSStandards/PHP_CodeSniffer/actions/runs/7986938143/job/21808386831?pr=355#step:9:56).

Since the sample size is very small, it is very possible that this variation in the execution time is not directly related to the changes in this branch.