@rodrigoprimo 2024-02-21
@jrfnl, now that I ran the tests a few more times, I'm starting to think that the first data that I got is not very reliable. I'm seeing a significant variation in execution times between runs. A few times, the execution time using this branch is faster than using the main branch and sometimes it is not.

I tried running all the tests without the first commit and didn't notice much of a difference (but again, the execution time varies a lot anyway). Then I used `--log-junit test-results.xml` and parsed the results to get the execution time for each of the new tests, and they don't seem particularly slow:

```
PHP_CodeSniffer\Tests\Core\Fixer\GenerateDiffTest.testGenerateDiffNoFile: 0.011799 seconds
PHP_CodeSniffer\Tests\Core\Fixer\GenerateDiffTest.testGenerateDiff with data set "whitespace diff at end of file": 0.002502 seconds
PHP_CodeSniffer\Tests\Core\Fixer\GenerateDiffTest.testGenerateDiff with data set "no difference": 0.002197 seconds
PHP_CodeSniffer\Tests\Core\Fixer\GenerateDiffTest.testGenerateDiffColoured: 0.002005 seconds
PHP_CodeSniffer\Tests\Core\Fixer\GenerateDiffTest.testGenerateDiffDifferentLineEndings: 0.001895 seconds
PHP_CodeSniffer\Tests\Core\Fixer\GenerateDiffTest.testGenerateDiff with data set "line removed": 0.001693 seconds
PHP_CodeSniffer\Tests\Core\Fixer\GenerateDiffTest.testGenerateDiff with data set "line added": 0.001689 seconds
PHP_CodeSniffer\Tests\Core\Fixer\GenerateDiffTest.testGenerateDiff with data set "blank lines at end of file": 0.001648 seconds
PHP_CodeSniffer\Tests\Core\Fixer\GenerateDiffTest.testGenerateDiff with data set "whitespace diff at start of file": 0.001544 seconds
PHP_CodeSniffer\Tests\Core\Fixer\GenerateDiffTest.testGenerateDiff with data set "blank lines at start of file": 0.001512 seconds
PHP_CodeSniffer\Tests\Core\Fixer\GenerateDiffTest.testGenerateDiff with data set "trailing whitespace removed": 0.001454 seconds
PHP_CodeSniffer\Tests\Core\Fixer\GenerateDiffTest.testGenerateDiff with data set "var name changed": 0.001442 seconds
PHP_CodeSniffer\Tests\Core\Fixer\GenerateDiffTest.testGenerateDiff with data set "tab replaced with spaces": 0.001417 seconds
```

For comparison, here are the slowest tests for this run:

```
PHP_CodeSniffer\Standards\Generic\Tests\WhiteSpace\ScopeIndentUnitTest.testSniff: 0.59299 seconds
PHP_CodeSniffer\Standards\Squiz\Tests\Arrays\ArrayDeclarationUnitTest.testSniff: 0.340927 seconds
PHP_CodeSniffer\Standards\Squiz\Tests\ControlStructures\ForLoopDeclarationUnitTest.testSniff: 0.224787 seconds
PHP_CodeSniffer\Standards\Squiz\Tests\Commenting\FunctionCommentUnitTest.testSniff: 0.186899 seconds
PHP_CodeSniffer\Standards\Squiz\Tests\WhiteSpace\OperatorSpacingUnitTest.testSniff: 0.15944 seconds
```