@jrfnl 2024-02-21
Just to give you some idea of the difference this makes on Windows (on a monster of a desktop) - when running just the GenerateDiff tests, I consistently see a difference of 2 seconds in the test run (~3 seconds without the patch), ~1 second with the patch):

Timing without the patch (first commit only):
```
testGenerateDiffNoFile: 0.240376 seconds
testGenerateDiff with data set "no difference": 0.240557 seconds
testGenerateDiff with data set "line removed": 0.226245 seconds
testGenerateDiff with data set "line added": 0.270723 seconds
testGenerateDiff with data set "var name changed": 0.226196 seconds
testGenerateDiff with data set "trailing whitespace removed": 0.226120 seconds
testGenerateDiff with data set "tab replaced with spaces": 0.226315 seconds
testGenerateDiff with data set "blank lines at start of file": 0.315879 seconds
testGenerateDiff with data set "whitespace diff at start of file": 0.226627 seconds
testGenerateDiff with data set "blank lines at end of file": 0.225380 seconds
testGenerateDiff with data set "whitespace diff at end of file": 0.240838 seconds
testGenerateDiffColoured: 0.226026 seconds
testGenerateDiffDifferentLineEndings: 0.226189 seconds
```

Timing with the patch:
```
testGenerateDiffNoFile: 0.077598 seconds
testGenerateDiff with data set "no difference": 0.075645 seconds
testGenerateDiff with data set "line removed": 0.076428 seconds
testGenerateDiff with data set "line added": 0.076219 seconds
testGenerateDiff with data set "var name changed": 0.074596 seconds
testGenerateDiff with data set "trailing whitespace removed": 0.075834 seconds
testGenerateDiff with data set "tab replaced with spaces": 0.075706 seconds
testGenerateDiff with data set "blank lines at start of file": 0.074717 seconds
testGenerateDiff with data set "whitespace diff at start of file": 0.075771 seconds
testGenerateDiff with data set "blank lines at end of file": 0.074711 seconds
testGenerateDiff with data set "whitespace diff at end of file": 0.075682 seconds
testGenerateDiffColoured: 0.075340 seconds
testGenerateDiffDifferentLineEndings: 0.075768 seconds
```