@rodrigoprimo 2024-02-21
I agree that there don't seem to be any performance regressions introduced by this PR.

Thanks for the additional details regarding the difference this PR makes in the execution time when running on Windows. I got curious and decided to check the execution times of the new tests without the patch on Linux, and there is an improvement as well.

Without the patch (just the first commit):

```
PHP_CodeSniffer\Tests\Core\Fixer\GenerateDiffTest.testGenerateDiffNoFile: 0.011705 seconds
PHP_CodeSniffer\Tests\Core\Fixer\GenerateDiffTest.testGenerateDiff with data set "tab replaced with spaces": 0.002201 seconds
PHP_CodeSniffer\Tests\Core\Fixer\GenerateDiffTest.testGenerateDiff with data set "blank lines at start of file": 0.002173 seconds
PHP_CodeSniffer\Tests\Core\Fixer\GenerateDiffTest.testGenerateDiff with data set "no difference": 0.00211 seconds
PHP_CodeSniffer\Tests\Core\Fixer\GenerateDiffTest.testGenerateDiff with data set "trailing whitespace removed": 0.001728 seconds
PHP_CodeSniffer\Tests\Core\Fixer\GenerateDiffTest.testGenerateDiff with data set "line added": 0.001649 seconds
PHP_CodeSniffer\Tests\Core\Fixer\GenerateDiffTest.testGenerateDiff with data set "line removed": 0.001616 seconds
PHP_CodeSniffer\Tests\Core\Fixer\GenerateDiffTest.testGenerateDiff with data set "blank lines at end of file": 0.001616 seconds
PHP_CodeSniffer\Tests\Core\Fixer\GenerateDiffTest.testGenerateDiff with data set "whitespace diff at start of file": 0.001575 seconds
PHP_CodeSniffer\Tests\Core\Fixer\GenerateDiffTest.testGenerateDiffDifferentLineEndings: 0.001547 seconds
PHP_CodeSniffer\Tests\Core\Fixer\GenerateDiffTest.testGenerateDiffColoured: 0.001528 seconds
PHP_CodeSniffer\Tests\Core\Fixer\GenerateDiffTest.testGenerateDiff with data set "whitespace diff at end of file": 0.001439 seconds
PHP_CodeSniffer\Tests\Core\Fixer\GenerateDiffTest.testGenerateDiff with data set "var name changed": 0.001392 seconds
```

With the patch:

```
PHP_CodeSniffer\Tests\Core\Fixer\GenerateDiffTest.testGenerateDiffNoFile: 0.002191 seconds
PHP_CodeSniffer\Tests\Core\Fixer\GenerateDiffTest.testGenerateDiff with data set "no difference": 0.001942 seconds
PHP_CodeSniffer\Tests\Core\Fixer\GenerateDiffTest.testGenerateDiff with data set "line removed": 0.001874 seconds
PHP_CodeSniffer\Tests\Core\Fixer\GenerateDiffTest.testGenerateDiff with data set "line added": 0.001625 seconds
PHP_CodeSniffer\Tests\Core\Fixer\GenerateDiffTest.testGenerateDiffColoured: 0.001599 seconds
PHP_CodeSniffer\Tests\Core\Fixer\GenerateDiffTest.testGenerateDiff with data set "blank lines at end of file": 0.001595 seconds
PHP_CodeSniffer\Tests\Core\Fixer\GenerateDiffTest.testGenerateDiff with data set "trailing whitespace removed": 0.001578 seconds
PHP_CodeSniffer\Tests\Core\Fixer\GenerateDiffTest.testGenerateDiff with data set "var name changed": 0.001555 seconds
PHP_CodeSniffer\Tests\Core\Fixer\GenerateDiffTest.testGenerateDiff with data set "whitespace diff at end of file": 0.001554 seconds
PHP_CodeSniffer\Tests\Core\Fixer\GenerateDiffTest.testGenerateDiff with data set "whitespace diff at start of file": 0.001546 seconds
PHP_CodeSniffer\Tests\Core\Fixer\GenerateDiffTest.testGenerateDiffDifferentLineEndings: 0.001503 seconds
PHP_CodeSniffer\Tests\Core\Fixer\GenerateDiffTest.testGenerateDiff with data set "tab replaced with spaces": 0.001345 seconds
PHP_CodeSniffer\Tests\Core\Fixer\GenerateDiffTest.testGenerateDiff with data set "blank lines at start of file": 0.001342 seconds
```